Final round of PHPCBF fixes (sans disabled rules)
