+++ components/Filesystem/Layer/class-chrootlayer.php
@@ -24,7 +24,7 @@ class ChrootLayer extends Layer {
 	 * @param  Filesystem $fs  The filesystem to chroot.
 	 * @param  string     $root  The root path to chroot to.
 	 */
-	function __construct( Filesystem $fs, $chroot ) {
+	public function __construct( Filesystem $fs, $chroot ) {
 		parent::__construct( $fs );
 		$this->chroot = $this->forward_slashes_on_local_filesystem_on_windows(
 			rtrim( $chroot, '/' ) . '/'
